ref: price 타입 bigDecimal로 변경

+++ product/src/main/java/com/smore/product/domain/entity/Product.java
@@ -3,6 +3,7 @@ package com.smore.product.domain.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.Clock;
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -29,7 +30,7 @@ public class Product {
     @Column(columnDefinition = "TEXT")
     private String description;
 
-    private int price;
+    private BigDecimal price;
 
     private int stock;
 
@@ -51,7 +52,7 @@ public class Product {
             UUID categoryId,
             String name,
             String description,
-            int price,
+            BigDecimal price,
             int stock,
             SaleType saleType,
             Integer thresholdForAuction
@@ -76,7 +77,7 @@ public class Product {
     public void update(
             String name,
             String description,
-            Integer price,
+            BigDecimal price,
             Integer stock,
             UUID categoryId,
             SaleType saleType,
@@ -107,7 +108,7 @@ public class Product {
         this.description = description;
     }
 
-    public void changePrice(Integer price) {
+    public void changePrice(BigDecimal price) {
         this.price = price;
     }
